+++ src/main/java/backend/user/User.java
@@ -1,4 +1,4 @@
-package backend;
+package backend.user;
 
 public class User {
     public static String userId;
