+++ src/main/java/frontend/Main.java
@@ -5,11 +5,11 @@ import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
 import javafx.stage.Stage;
 
-public class Signup extends Application {
+public class Main extends Application {
 
     @Override
     public void start(Stage primaryStage) throws Exception {
-        FXMLLoader fxmlLoader = new FXMLLoader(Signup.class.getResource("signup.fxml"));
+        FXMLLoader fxmlLoader = new FXMLLoader(Main.class.getResource("signup.fxml"));
         Scene scene = new Scene(fxmlLoader.load(),800,700);
         primaryStage.setTitle("Signup");
         primaryStage.setScene(scene);
