+++ src/main/java/module-info.java
@@ -8,4 +8,6 @@ module com.example.connecthub {
 
     opens frontend to javafx.fxml;
     exports frontend;
+    exports frontend.userManagement;
+    opens frontend.userManagement to javafx.fxml;
 }
